+++ tests/TestCase/Shell/UserShellTest.php
@@ -12,16 +12,14 @@
  */
 namespace MeCms\Test\TestCase\Shell;
 
-use Cake\Console\ConsoleIo;
 use Cake\ORM\TableRegistry;
-use Cake\TestSuite\Stub\ConsoleOutput;
 use MeCms\Shell\UserShell;
-use MeTools\TestSuite\TestCase;
+use MeTools\TestSuite\ConsoleIntegrationTestCase;
 
 /**
  * InstallShellTest class
  */
-class UserShellTest extends TestCase
+class UserShellTest extends ConsoleIntegrationTestCase
 {
     /**
      * @var \MeCms\Model\Table\UsersTable
@@ -43,38 +41,6 @@ class UserShellTest extends TestCase
     ];
 
     /**
-     * Internal method.
-     * It sets some expectations for shell on `in()` method.
-     * @return void
-     */
-    protected function setShellExpectsForInMethod()
-    {
-        $this->UserShell->expects($this->at(0))
-            ->method('in')
-            ->will($this->returnValue('myusername'));
-
-        $this->UserShell->expects($this->at(1))
-            ->method('in')
-            ->will($this->returnValue('password1/'));
-
-        $this->UserShell->expects($this->at(2))
-            ->method('in')
-            ->will($this->returnValue('password1/'));
-
-        $this->UserShell->expects($this->at(3))
-            ->method('in')
-            ->will($this->returnValue('myemail@example.com'));
-
-        $this->UserShell->expects($this->at(4))
-            ->method('in')
-            ->will($this->returnValue('Alfa'));
-
-        $this->UserShell->expects($this->at(5))
-            ->method('in')
-            ->will($this->returnValue('Beta'));
-    }
-
-    /**
      * Setup the test case, backup the static object values so they can be
      * restored. Specifically backs up the contents of Configure and paths in
      *  App if they have not already been backed up
@@ -86,17 +52,7 @@ class UserShellTest extends TestCase
 
         $this->Users = TableRegistry::get(ME_CMS . '.Users');
 
-        $this->out = new ConsoleOutput;
-        $this->err = new ConsoleOutput;
-        $this->io = new ConsoleIo($this->out, $this->err);
-        $this->io->level(2);
-
-        $this->UserShell = $this->getMockBuilder(UserShell::class)
-            ->setMethods(['in', '_stop'])
-            ->setConstructorArgs([$this->io])
-            ->getMock();
-
-        $this->UserShell->initialize();
+        $this->UserShell = new UserShell;
     }
 
     /**
@@ -105,130 +61,67 @@ class UserShellTest extends TestCase
      */
     public function testAdd()
     {
-        $this->setShellExpectsForInMethod();
+        $example = ['myusername', 'password1/', 'password1/', 'mail@example.com', 'Alfa', 'Beta'];
+        $id = 1 + $this->Users->find()->order(['id' => 'DESC'])->extract('id')->first();
 
-        $this->UserShell->expects($this->at(6))
-            ->method('in')
-            ->will($this->returnValue(3));
+        $this->exec('me_cms.user add', array_merge($example, ['3']));
+        $this->assertExitWithSuccess();
+        $this->assertOutputContains('<question>Group ID</question>');
+        $this->assertOutputContains('<success>The operation has been performed correctly</success>');
+        $this->assertOutputContains('<success>The user was created with ID ' . $id . '</success>');
 
-        $id = $this->UserShell->add();
-        $this->assertNotEmpty($id);
-
-        $this->assertEquals([
-            '+----+---------+',
-            '| <info>ID</info> | <info>Name</info>    |',
-            '+----+---------+',
-            '| 1  | Admin   |',
-            '| 4  | Fans    |',
-            '| 2  | Manager |',
-            '| 5  | People  |',
-            '| 3  | User    |',
-            '+----+---------+',
-            '<success>' . I18N_OPERATION_OK . '</success>',
-        ], $this->out->messages());
-        $this->assertEmpty($this->err->messages());
-
-        $user = $this->Users->find()->where(['id' => $id])->first();
-
+        //Checks the user has been created
+        $user = $this->Users->find()->where(compact('id'))->first();
         $this->assertNotEmpty($user);
         $this->assertEquals(3, $user->group_id);
-    }
 
-    /**
-     * Test for `add()` method, with invalid values
-     * @test
-     */
-    public function testAddInvalidValues()
-    {
-        $this->UserShell->expects($this->any())
-            ->method('in')
-            ->will($this->returnValue('test'));
+        $this->Users->delete($this->Users->get($id));
+        $id++;
 
-        $this->UserShell->params['group'] = 2;
-        $this->UserShell->params['verbose'] = true;
-        $this->assertFalse($this->UserShell->add());
+        //Tries using the `group` param
+        $this->exec('me_cms.user add --group 2', $example);
+        $this->assertExitWithSuccess();
+        $this->assertOutputContains('<success>The operation has been performed correctly</success>');
+        $this->assertOutputContains('<success>The user was created with ID ' . $id . '</success>');
 
-        $this->assertEmpty($this->out->messages());
-        $this->assertEquals([
-            '<error>' . I18N_OPERATION_NOT_OK . '</error>',
-            '<error>The user could not be saved</error>',
-            '<error>Field `email`: you have to enter a valid value</error>',
-            '<error>Field `first_name`: allowed chars: letters, apostrophe, space. Has to begin with a capital letter</error>',
-            '<error>Field `last_name`: allowed chars: letters, apostrophe, space. Has to begin with a capital letter</error>',
-            '<error>Field `password`: must be at least 8 chars</error>',
-        ], $this->err->messages());
-    }
-
-    /**
-     * Test for `add()` method, with no users groups
-     * @test
-     */
-    public function testAddNoUsersGroups()
-    {
-        //Deletes all groups
-        $this->assertNotEquals(0, $this->Users->Groups->deleteAll(['id >=' => '1']));
-
-        $this->assertFalse($this->UserShell->add());
-
-        $this->assertEmpty($this->out->messages());
-        $this->assertEquals([
-            '<error>Before you can manage users, you have to create at least a user group</error>',
-        ], $this->err->messages());
-    }
-
-    /**
-     * Test for `add()` method, using the `group` param
-     * @test
-     */
-    public function testAddUsingGroupParam()
-    {
-        $this->setShellExpectsForInMethod();
-
-        $this->UserShell->params['group'] = 2;
-        $id = $this->UserShell->add();
-        $this->assertNotEmpty($id);
-
-        $this->assertEquals(['<success>' . I18N_OPERATION_OK . '</success>'], $this->out->messages());
-        $this->assertEmpty($this->err->messages());
-
-        $user = $this->Users->find()->where(['id' => $id])->first();
-
+        //Checks the user has been created
+        $user = $this->Users->find()->where(compact('id'))->first();
         $this->assertNotEmpty($user);
         $this->assertEquals(2, $user->group_id);
-    }
 
-    /**
-     * Test for `add()` method, with empty values
-     * @test
-     */
-    public function testAddWithEmptyValues()
-    {
-        $this->UserShell->params['group'] = 2;
-        $this->assertFalse($this->UserShell->add());
+        //Tries with a no existing group
+        $this->exec('me_cms.user add --group 123', $example);
+        $this->assertExitWithError();
+        $this->assertErrorContains('<error>Invalid group ID</error>');
 
-        $this->assertEmpty($this->out->messages());
-        $this->assertEquals([
-            '<error>Field `username` is empty. Try again</error>',
-        ], $this->err->messages());
-    }
+        //Tries with empty data
+        $this->exec('me_cms.user add -v', []);
+        $this->assertExitWithError();
+        $this->assertErrorContains('<error>Field `username` is empty. Try again</error>');
 
-    /**
-     * Test for `add()` method, with a no existing group
-     * @test
-     */
-    public function testAddWithNoExistingGroup()
-    {
-        $this->UserShell->expects($this->any())
-            ->method('in')
-            ->will($this->returnValue('test'));
+        //Tries with wrong data
+        $this->exec('me_cms.user add -v', ['ab', 'password', 'password2', 'mail', 'aa', 'bb', '3']);
+        $this->assertExitWithError();
+        $this->assertErrorContains('<error>The operation has not been performed correctly</error>');
+        $this->assertErrorContains('<error>The user could not be saved</error>');
+        $this->assertErrorContains('<error>Field `email`: you have to enter a valid value</error>');
+        $this->assertErrorContains('<error>Field `first_name`: must be between 3 and 40 chars</error>');
+        $this->assertErrorContains('<error>Field `first_name`: allowed chars: letters, apostrophe, ' .
+            'space. Has to begin with a capital letter</error>');
+        $this->assertErrorContains('<error>Field `last_name`: must be between 3 and 40 chars</error>');
+        $this->assertErrorContains('<error>Field `last_name`: allowed chars: letters, apostrophe, ' .
+            'space. Has to begin with a capital letter</error>');
+        $this->assertErrorContains('<error>Field `username`: must be between 4 and 40 chars</error>');
+        $this->assertErrorContains('<error>Field `username`: allowed chars: lowercase letters, numbers, dash</error>');
+        $this->assertErrorContains('<error>Field `password`: the password should contain letters, ' .
+            'numbers and symbols</error>');
+        $this->assertErrorContains('<error>Field `password_repeat`: passwords don\'t match</error>');
 
-        $this->UserShell->params['group'] = 123;
-        $this->assertFalse($this->UserShell->add());
-
-        $this->assertEmpty($this->out->messages());
-        $this->assertEquals([
-            '<error>Invalid group ID</error>',
-        ], $this->err->messages());
+        //Tries with no groups
+        $this->assertNotEquals(0, $this->Users->Groups->deleteAll(['id >=' => '1']));
+        $this->exec('me_cms.user add -v');
+        $this->assertExitWithError();
+        $this->assertErrorContains('<error>Before you can manage users, you have to create at least a user group</error>');
     }
 
     /**
@@ -237,13 +130,8 @@ class UserShellTest extends TestCase
      */
     public function testGroups()
     {
-        $this->UserShell->groups();
-
-        //Deletes all groups
-        $this->assertNotEquals(0, $this->Users->Groups->deleteAll(['id >=' => '1']));
-
-        $this->UserShell->groups();
-
+        $this->exec('me_cms.user groups');
+        $this->assertExitWithSuccess();
         $this->assertEquals([
             '+----+---------+---------+-------+',
             '| <info>ID</info> | <info>Name</info>    | <info>Label</info>   | <info>Users</info> |',
@@ -254,8 +142,14 @@ class UserShellTest extends TestCase
             '| 4  | fans    | Fans    | 3     |',
             '| 5  | people  | People  | 0     |',
             '+----+---------+---------+-------+',
-        ], $this->out->messages());
-        $this->assertEquals(['<error>There are no user groups</error>'], $this->err->messages());
+        ], $this->_out->messages());
+
+        //Deletes all groups
+        $this->assertNotEquals(0, $this->Users->Groups->deleteAll(['id >=' => '1']));
+
+        $this->exec('me_cms.user groups');
+        $this->assertExitWithError();
+        $this->assertErrorContains('<error>There are no user groups</error>');
     }
 
     /**
@@ -264,16 +158,9 @@ class UserShellTest extends TestCase
      */
     public function testUsers()
     {
-        $this->UserShell->users();
-
-        $users = TableRegistry::get(ME_CMS . '.Users');
-
-        //Deletes all users
-        $this->assertNotEquals(0, $users->deleteAll(['id >=' => '1']));
-
-        $this->UserShell->users();
-
-        $this->assertTextEquals([
+        $this->exec('me_cms.user users');
+        $this->assertExitWithSuccess();
+        $this->assertEquals([
             '+----+----------+-------+--------------+-------------------+-------+---------+------------------+',
             '| <info>ID</info> | <info>Username</info> | <info>Group</info> | <info>Name</info>         | <info>Email</info>             | <info>Posts</info> | <info>Status</info>  | <info>Date</info>             |',
             '+----+----------+-------+--------------+-------------------+-------+---------+------------------+',
@@ -283,8 +170,14 @@ class UserShellTest extends TestCase
             '| 4  | abc      | User  | Abc Def      | abc@example.com   | 1     | Active  | 2016/12/24 17:03 |',
             '| 5  | delta    | Admin | Mno Pqr      | delta@example.com | 0     | Active  | 2016/12/24 17:04 |',
             '+----+----------+-------+--------------+-------------------+-------+---------+------------------+',
-        ], $this->out->messages());
-        $this->assertEquals(['<error>There are no users</error>'], $this->err->messages());
+        ], $this->_out->messages());
+
+        //Deletes all users
+        $this->assertNotEquals(0, $this->Users->deleteAll(['id >=' => '1']));
+
+        $this->exec('me_cms.user users');
+        $this->assertExitWithError();
+        $this->assertErrorContains('<error>There are no users</error>');
     }
 
     /**