+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### 2.21.0
 * updated for CakePHP 3.5;
 * `UserShell::add()` now returns `true`. The created user id is shown through a
-    successful message.
+    successful message;
+* uses `ConsoleIntegrationTestCase`. Console tests have been simplified.
 
 ## 2.20 branch
 ### 2.20.2